EVEREST-451 improved e2e tests
